feat: support to decode array and map
Add GetVariantArray to parse the value as an array.
All nodes in the array have individual tag.

Add GetArray to parse the value as an array.
All nodes in the array have the same tag.

Add GetStringMap to parse the value as an key-value pair,
and the key type is string.